requirements.txt
pillow
numpy
scipy
matplotlib
joblib
tqdm